T: OH7: merging bug-fixes from trunk

+++ branches/OH7_dev/CTablemap/CTablemap.cpp
@@ -1,15 +1,15 @@
-//******************************************************************************
+//*****************************************************************************
 //
 // This file is part of the OpenHoldem project
 //   Download page:         http://code.google.com/p/openholdembot/
 //   Forums:                http://www.maxinmontreal.com/forums/index.php
 //   Licensed under GPL v3: http://www.gnu.org/licenses/gpl.html
 //
-//******************************************************************************
+//*****************************************************************************
 //
 // Purpose:
 //
-//******************************************************************************
+//*****************************************************************************
 
 
 #include "StdAfx.h"
@@ -672,7 +672,7 @@ int CTablemap::LoadTablemap(const CString _fname)
 			hold_image.image = new RGBAImage(hold_image.width, hold_image.height, t.GetString());
 
 			// read next "height" lines
-			for (y=0; y < (int) hold_image.height; y++) 
+			for (y=0; y < hold_image.height; y++) 
 			{
 				linenum++;
 				if (!ar.ReadString(strLine)) 
@@ -681,7 +681,7 @@ int CTablemap::LoadTablemap(const CString _fname)
 					return ERR_SYNTAX;
 				}
 				// scan across "width" of line to get values
-				for (x=0; x < (int) hold_image.width; x++) 
+				for (x=0; x < hold_image.width; x++) 
 				{
 					// unreverse bgra to abgr
 					hexval = strLine.Mid(x*8+6, 2) + strLine.Mid(x*8, 6);
